✨ (feat): Add a field for inputting user answers and alert prompts to display user progress
Added a TextField for user to input their answers and validation checks against the answer value of the Question struct. Switched the TextField keyboard to use decimalPad along with toolbar buttons to submit the user's answer or to cancel the focus on the TextField View. Also added alert prompts to show the user progress each time the user submits their answer.