✨ (feat): Add ability for users to start a new game round
Added a new @State property endGame to watch for the end of the current game round and a new method resetGame that resets the game to its initial state. An additional alert prompt has been added to show users their summary of game progress with how many answers they got correct and two buttons in the prompt for users to choose either to start a new round or simply dismiss the prompt.